+++ R/tool_misc.R
@@ -11,12 +11,12 @@ bdiag <- function(...){
   else
     x <- list(...)
   n <- length(x)
-  if(n == 0) return(NULL)
+  if(n == 0L) return(NULL)
   x <- lapply(x, function(y) if(length(y)) as.matrix(y) else
               stop("Zero-length component in x"))
   d <- array(unlist(lapply(x, dim)), c(2, n))
-  rr <- d[1, ]
-  cc <- d[2, ]
+  rr <- d[1L, ]
+  cc <- d[2L, ]
   rsum <- sum(rr)
   csum <- sum(cc)
   out <- array(0, c(rsum, csum))
@@ -29,7 +29,7 @@ bdiag <- function(...){
   ind[4,   ] <- ccum
   imat <- array(1:(rsum * csum), c(rsum, csum))
   iuse <- apply(ind, 2, function(y, imat) imat[(y[1L]+1):y[2L],
-                                               (y[3L]+1):y[4L]], imat=imat)
+                                               (y[3L]+1):y[4L]], imat = imat)
   iuse <- as.vector(unlist(iuse))
   out[iuse] <- unlist(x)
   return(out)
@@ -157,15 +157,15 @@ pres <- function(x) {  # pres.panelmodel
   ## but used in residuals.pggls, residuals.pcce, residuals.pmg
   
   ## extract indices
-  groupind <-attr(x$model, "index")[ , 1L]
-  timeind  <-attr(x$model, "index")[ , 2L]
+  groupind <- attr(x$model, "index")[ , 1L]
+  timeind  <- attr(x$model, "index")[ , 2L]
   
   # fix to allow operation with pggls, pmg
   # [TODO: one day, make this cleaner; with the describe framework?]
   if (!is.null(x$args$model))                 maybe_fd <- x$args$model
   if (!is.null(attr(x, "pmodel")$model.name)) maybe_fd <- attr(x, "pmodel")$model.name # this line is currently needed to detect pggls models
   
-  ## Achim's fix: reduce id and time index to accomodate first-differences model's number of observations
+  ## Achim's fix: reduce id and time index to accommodate first-differences model's number of observations
   if(exists("maybe_fd") && maybe_fd == "fd") {
     groupi <- as.numeric(groupind)
     ## make vector =1 on first obs in each group, 0 elsewhere
@@ -183,10 +183,15 @@ pres <- function(x) {  # pres.panelmodel
 }
 
 
-# punbalancedness: measures for unbalancedness of a pandel data set as
-# defined in \insertCite{AHRE:PINC:81;textual}{plm}, p. 228 (gamma and
-# nu) and for nested panel structures as in
-# \insertCite{BALT:SONG:JUNG:01;textual}{plm}, pp. 368-369
+# punbalancedness: measures for unbalancedness of a panel data set as
+# defined in Ahrens/Pincus (1981), p. 228 (gamma and
+# nu) and for nested panel structures as in Baltagi/Song/Jung (2001), pp. 368-369 .
+#
+# Ahrens/Pincus (1981), On Two Measures of Unbalancedness in a One-Way Model
+#  and Their Relation to Efficiency, Biometrical Journal, Vol. 23, pp. 227-235.
+# 
+# Baltagi/Song/Jung (2001), The unbalanced nested error component regression model,
+#  Journal of Econometrics, Vol. 101, pp. 357-381
 
 
 #' Measures for Unbalancedness of Panel Data
@@ -316,6 +321,7 @@ punbalancedness <- function(x, ...) {
 punbalancedness.default <- function(x, ...) {
 
   ii <- index(x)
+  if(!is.index(ii)) stop("no valid index found for input object 'x'")
   
   if (ncol(ii) == 2L) {
    ## original Ahrens/Pincus (1981)
@@ -461,7 +467,7 @@ pvar.default <- function(x, id, time, ...){
   id.variation   <- rep(TRUE, len)
   time.variation_anyNA <- rep(FALSE, len)
   id.variation_anyNA   <- rep(FALSE, len)
-  lid <- split(x, id)   # these split() functions seem particularly slow
+  lid   <- split(x, id)   # these split() functions seem particularly slow
   ltime <- split(x, time)
   if(is.list(x)){
     if(len == 1L){
@@ -545,7 +551,7 @@ pvar.data.frame <- function(x, index = NULL, ...){
 #' @export
 pvar.pdata.frame <- function(x, ...){
   index <- attr(x, "index")
-  id <- index[[1L]]
+  id   <- index[[1L]]
   time <- index[[2L]]
   pvar.default(x, id, time)
 }
@@ -566,7 +572,7 @@ print.pvar <- function(x, ...){
   if(any(!x$time.variation)){
     var <- varnames[x$time.variation == FALSE]
     #    if (!is.null(y)) var <- var[-which(var==y$id)]
-    if (length(var)!=0) cat(paste("no time variation:      ", paste(var,collapse=" "),"\n"))
+    if(length(var)!=0) cat(paste("no time variation:      ", paste(var,collapse=" "),"\n"))
   }
   if(any(!x$id.variation)){
     var <- varnames[x$id.variation == FALSE]
@@ -577,39 +583,12 @@ print.pvar <- function(x, ...){
   # any individual-time combinations all NA?
   if(any(x$time.variation_anyNA)){
     var_anyNA <- varnames[x$time.variation_anyNA]
-    if (length(var_anyNA)!=0) cat(paste("all NA in time dimension for at least one individual: ", paste(var_anyNA,collapse=" "),"\n"))
+    if(length(var_anyNA)!=0) cat(paste("all NA in time dimension for at least one individual: ", paste(var_anyNA,collapse=" "),"\n"))
   }
   if(any(x$id.variation_anyNA)){
     var_anyNA <- varnames[x$id.variation_anyNA]
-    if (length(var_anyNA)!=0) cat(paste("all NA in ind. dimension for at least one time period:", paste(var_anyNA,collapse=" "),"\n"))
+    if(length(var_anyNA)!=0) cat(paste("all NA in ind. dimension for at least one time period:", paste(var_anyNA,collapse=" "),"\n"))
   }
 }
 
-## Non-exported internal function for subsetting of pseries. Can be used
-## internally. 
-## Currently we do not have "proper" subsetting function for pseries
-## ([.pseries) which applies in any case (but see the sketch to be tested in
-##  tool_pdata.frame.R)
-subset_pseries <- function(x, ...) {
 
- ## use '...' instead of only one specific argument, because subsetting for
- ## factors can have argument 'drop', e.g., x[i, drop=TRUE] see ?Extract.factor
-  index <- attr(x, "index")
-  if (is.null(index)) warning("pseries object with is.null(index(pseries)) == TRUE encountered, trying to continue anyway...")
-  if (!is.index(index)) stop(paste0("pseries object has illegal index with class(index) == ", paste0(class(index), collapse = ", ")))
-  names_orig <- names(x)
-  x <- remove_pseries_features(x)
-  result <- x[...]
-
-  # subset index / identify rows to keep in the index:
-  keep_rownr <- seq_along(names_orig)  # full length row numbers original pseries
-  names(keep_rownr) <- names_orig
-  keep_rownr <- keep_rownr[names(result)] # row numbers to keep after subsetting
-  index <- index[keep_rownr, ]
-
-  # drop unused levels (like in subsetting of pdata.frames)
-  index <- droplevels(index)
-
-  result <- add_pseries_features(result, index)
-  return(result)
-}